move IntReference
